requirements.txt
clickhouse-driver>=0.2.5
pandas>=1.3.0
PyYAML>=6.0
numpy>=1.21.0
python-dateutil>=2.8.2
pytz>=2021.1
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
ipython>=7.31.0
neo4j>=5.8.0
networkx>=3.1.0  # 用于图数据处理和可视化 